feat: tosspayments 결제v1-1 #53

+++ backend/src/main/java/com/malnutrition/backend/domain/lecture/lecture/dto/LectureRequestDto.java
@@ -13,7 +13,7 @@ public class LectureRequestDto {
 
     private String title;      // 강의 제목
     private String content;    // 강의 내용
-    private Integer price;     // 강의 가격
+    private Long price;     // 강의 가격
     private String lectureLevel; // 강의 수준 (초급, 중급, 고급 등)
     private String lectureStatus; // 강의 상태 (OPEN, CLOSED 등)
